Split preproc into several files
These files are shorter, and thus easier to navigate than a big monolith.
